Annotations= how the code is working. For V2.
V2 was first code which solved all easy-hard-harder problems. later on versions are modification of that. The proposed modifications are also annotated. So, this can be used to understand roughly that how the code is working. I had used my logic, and help of chatGPT (in writing code part). The total time taken to annotate was 12+ hours. The total time given in development of all versions is 1 week+.